Move globaltypes
