+++ config/globalTypes/global-state.d.ts
