+++ config/globalTypes/router-types.d.ts
